refactor(frontend/src/features/commands/list
/Command.jsx):

move Command.jsx to features/command/list

for better feature organization
